Strategic reformulation
Some classes were reorganized in different modules. Its shall allow a better implementation of the heterogeneous rocket model.
